requirements.txt
loguru==0.7.2
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.36
asyncpg==0.30.0
transitions==0.9.2
vk_api==11.9.9
